[FEAT] implement complete logic for registering bill information
