+++ hamso/src/main/java/com/halo/hamso/common/exception/TimeDuplicateException.java
@@ -1,0 +1,9 @@
+package com.halo.hamso.common.exception;
+
+import java.time.LocalTime;
+
+public class TimeDuplicateException extends RuntimeException{
+    public TimeDuplicateException(String message) {
+        super(message);
+    }
+}
